Refactor elf_loader into src/io/ with PIMPL and file splitting
Split the monolithic 2100-line elf_loader.cpp into focused modules under
src/io/:
- elf_loader.hpp: Public PIMPL header hiding elfio/libbtf types
- elf_reader.hpp: Internal shared types and declarations
- elf_loader.cpp: ElfObject facade with cached ELF reader
- elf_reader.cpp: ProgramReader + read_elf + utility functions
- elf_map_parser.cpp: Map/BTF/global-variable extraction
- elf_core_reloc.cpp: CO-RE relocation engine
- elf_extern_resolve.cpp: External symbol resolution

Key improvements:
- ElfObject uses PIMPL (ElfObjectState) to hide implementation types
- Cached ELFIO reader avoids re-parsing the file per section load
- Indexed function relocation lookup (set instead of linear scan)
