+++ src/ebpf_verifier.hpp
@@ -3,7 +3,7 @@
 #pragma once
 
 #include "config.hpp"
-#include "elf_loader.hpp"
+#include "io/elf_loader.hpp"
 #include "ir/program.hpp"
 #include "ir/unmarshal.hpp"
 #include "platform.hpp"
