+++ src/io/elf_loader.hpp
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
-#include <istream>
+#include <memory>
 #include <stdexcept>
 #include <string>
-#include <unordered_map>
 #include <vector>
 
 #include "ir/program.hpp"
 #include "platform.hpp"
+
 namespace prevail {
 
 class UnmarshalError : public std::runtime_error {
@@ -35,27 +35,13 @@ struct ElfProgramInfo {
     std::string invalid_reason;
 };
 
+/// @note Not thread-safe. Callers must synchronize externally.
 class ElfObject {
-    struct QueryKey {
-        std::string section;
-        std::string program;
-
-        bool operator==(const QueryKey&) const = default;
-    };
-
-    struct SectionCacheEntry {
-        bool loaded{};
-        bool valid{};
-        std::string error;
-        std::vector<RawProgram> programs;
-    };
-
-    struct QueryKeyHash {
-        size_t operator()(const QueryKey& key) const noexcept;
-    };
-
   public:
     ElfObject(std::string path, ebpf_verifier_options_t options, const ebpf_platform_t* platform);
+    ~ElfObject();
+    ElfObject(ElfObject&&) noexcept;
+    ElfObject& operator=(ElfObject&&) noexcept;
 
     const std::string& path() const noexcept;
     const std::vector<ElfProgramInfo>& list_programs();
@@ -64,22 +50,8 @@ class ElfObject {
     static bool is_valid(const ElfProgramInfo& program) noexcept { return !program.invalid; }
 
   private:
-    void discover_programs();
-    void load_section(const std::string& section_name);
-    std::vector<RawProgram> filter_section_programs(const std::vector<RawProgram>& programs,
-                                                    const std::string& desired_program) const;
-    void mark_section_validity(const std::string& section_name, bool valid, const std::string& reason);
-
-    std::string path_;
-    ebpf_verifier_options_t options_;
-    const ebpf_platform_t* platform_;
-
-    bool catalog_loaded_{};
-    std::vector<ElfProgramInfo> programs_;
-    std::vector<std::string> section_order_;
-    std::unordered_map<std::string, std::vector<size_t>> section_program_indices_;
-    std::unordered_map<std::string, SectionCacheEntry> section_cache_;
-    std::unordered_map<QueryKey, std::vector<RawProgram>, QueryKeyHash> query_cache_;
+    struct ElfObjectState;
+    std::unique_ptr<ElfObjectState> state_;
 };
 
 } // namespace prevail
